packages/middlewares/text-processors/newspaper3k/src/lambdas/parser/requirements.txt
boto3==1.34.34
aws_lambda_powertools==2.33.0
aws-xray-sdk==2.12.1
newspaper3k==0.2.8
nltk==3.8.1
regex==2023.12.25
